Close figures in each test notebook cell
These are normally closed by the matplotlib-inline backend, if you are
running the notebook itself, so there's no need for explicit closing.
However, the tests `exec` each cell in a non-notebook process, so this
automatic closing does not occur and needs to be done explicitly. Doing
so prevents a "too many figures have been opened" warning from
Matplotlib.